// src/model/bookings/bookings.model.ts
export interface Bookings {
  key?: string;
  Date: string;
  startTime: string;
  // endTime: string;
  Duration: string;
  Pickup: string;
  Destination: string;
  Assistance: string;
  EscortsGender: string;
  Carpool: string;
  Image: string;
  // Contact: number;
  Status: string;
  Driver: string;
  ROD: string;
  Payment: string;
  Comments: string;
  Email: string;
}